state deltas refactor
